server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["Timur Safin <timurtlt96@mail.ru>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.70.0"
uvicorn = "^0.15.0"
gunicorn = "^20.1.0"
pydantic = "^1.8.2"
SQLAlchemy = "^1.4.27"
psycopg2-binary = "^2.9.2"
SQLAlchemy-Utils = "^0.37.9"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
requests = "^2.26.0"
factory-boy = "^3.2.1"
freezegun = "^1.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "-vv -s -p no:warnings"
testpaths = [
    "tests",
]
